requirements.txt
chromadb == 0.6.3
pandas == 2.2.3 
pymilvus == 2.5.5
langchain_openai == 0.3.8
langchain == 0.3.20 
pypdf == 5.3.1
google-genai==1.5.0
langchain-community==0.3.19